selftests: cgroup: Increase timeout for test_cpu and test_kmem
We found that on some KVM instances and bare-metal systems, the following
cgroup selftests take longer than the kselftest framework's default timeout
of 45 seconds to complete. As a result, they are being terminated with a
TIMEOUT error:

  * selftests: cgroup: test_cpu  # TIMEOUT at 45 seconds
  * selftests: cgroup: test_kmem # TIMEOUT at 45 seconds

This commit increases the timeout for these tests to ensure they have enough
time to complete successfully on slower systems.

Signed-off-by: Naresh Kamboju <naresh.kamboju@linaro.org>